Edit results section
I changed the order, per the ordering of the rest of the paper. I
also removed the log-compression part of the ASSR experiment. The
advantage of the log is already in a sense established by the ASC
experiments, so we can gain some space (and repeat ourselves a
little less) but removing this part. Of course, we can always put
it back in.